Refactor CoffeeTasteScanner helpers
