+++ src/screens/CoffeeTasteScanner/CoffeeTasteScanner.tsx
@@ -57,14 +57,54 @@ import type {
   ConfirmStructuredPayload,
   CoffeeEvaluationResult,
 } from './services';
-import { BrewContext } from '../../types/Personalization';
-import type { TasteProfileVector, UserTasteProfile } from '../../types/Personalization';
 import { usePersonalization } from '../../hooks/usePersonalization';
 import { showToast } from '../../utils/toast';
-import { buildScanPreferenceComparison } from '../../utils/scanPreferenceComparison';
 import { API_URL } from '../../services/api';
 import { recognizeCoffee } from 'services/VisionService.ts';
-import { parseAIResponse } from '../../components/utils/AITextFormatter.ts';
+import {
+  buildComparisonText,
+  extractSentenceBlocks,
+  resolveInsightHeadline,
+  resolveVerdictExplanationText,
+} from './comparison';
+import {
+  COFFEE_GRADIENT,
+  IMAGE_QUALITY,
+  MAX_BASE64_LENGTH,
+  MAX_IMAGE_DIMENSION,
+  STRUCTURED_FIELD_LABELS,
+  STRUCTURED_FIELD_ORDER,
+  WELCOME_GRADIENT,
+} from './constants';
+import { extractPreferenceSnapshot } from './preference';
+import type { CoffeePreferenceSnapshot } from './preference';
+import {
+  createStructuredFieldsFromMetadata,
+  isStructuredValueFilled,
+  normalizeStructuredStringArrayValue,
+  normalizeStructuredStringValue,
+  structuredFieldsToConfidence,
+  structuredFieldsToMetadata,
+} from './structuredFields';
+import {
+  clampTasteValue,
+  extractTasteVectorFromPayload,
+  normalizeTasteVectorTo10,
+  resolveTasteProfile,
+} from './tasteProfile';
+import type {
+  StructuredFieldKey,
+  StructuredFieldsState,
+  StructuredListFieldKey,
+  StructuredTextFieldKey,
+} from './types';
+import {
+  buildBrewContext,
+  ensureFileUri,
+  isOfflineError,
+  normalizeConfidenceScore,
+  stripFileUri,
+} from './utils';
 
 interface ScanResult {
   original: string;
@@ -90,14 +130,6 @@ interface ScanResult {
 
 type ScanResultLike = ScanResult & { rawStructuredResponse?: unknown };
 
-type CoffeePreferenceSnapshot = {
-  ai_recommendation?: string | null;
-  consistency_score?: number | null;
-  taste_vector: Record<string, number> | null;
-  updatedAt?: string | null;
-  lastRecalculatedAt?: string | null;
-};
-
 type StructuredConfirmPayload = ConfirmStructuredPayload & {
   correctedText?: string | null;
   purchased?: boolean | null;
@@ -108,703 +140,7 @@ interface ProfessionalOCRScannerProps {
   onHistoryPress?: () => void;
   onQuestionnairePress?: () => void;
 }
-
-const resolveTimeOfDay = (date: Date): BrewContext['timeOfDay'] => {
-  const hour = date.getHours();
-  if (hour < 12) {
-    return 'morning';
-  }
-  if (hour < 18) {
-    return 'afternoon';
-  }
-  if (hour < 22) {
-    return 'evening';
-  }
-  return 'night';
-};
-
-const getIsoWeekday = (date: Date): number => {
-  const weekday = date.getDay();
-  return weekday === 0 ? 7 : weekday;
-};
-
-const buildBrewContext = (metadata?: Record<string, unknown>): BrewContext => {
-  const now = new Date();
-  const context: BrewContext = {
-    timeOfDay: resolveTimeOfDay(now),
-    weekday: getIsoWeekday(now),
-  };
-
-  if (metadata && Object.keys(metadata).length > 0) {
-    context.metadata = { ...metadata };
-  }
-
-  return context;
-};
-
-const isOfflineError = (error: unknown): boolean => {
-  if (!(error instanceof Error)) {
-    return false;
-  }
-  return error.message === 'Offline' || error.message.includes('Network request failed');
-};
-
-const ensureFileUri = (path: string): string => {
-  if (path.startsWith('file://')) {
-    return path;
-  }
-  return `file://${path}`;
-};
-
-const stripFileUri = (path: string): string => {
-  if (path.startsWith('file://')) {
-    return path.replace('file://', '');
-  }
-  return path;
-};
-
-const extractPreferenceSnapshot = (
-  payload?: Record<string, unknown> | null,
-): CoffeePreferenceSnapshot | null => {
-  if (!payload) {
-    return null;
-  }
-  const rawCoffeePreferences =
-    payload.coffee_preferences && typeof payload.coffee_preferences === 'object'
-      ? (payload.coffee_preferences as Record<string, unknown>)
-      : null;
-  const coffeePreferences = rawCoffeePreferences ?? payload;
-  const aiRecommendation =
-    typeof payload.ai_recommendation === 'string' ? payload.ai_recommendation : null;
-  const consistencyScore =
-    typeof payload.consistency_score === 'number'
-      ? payload.consistency_score
-      : typeof payload.ai_confidence === 'number'
-        ? payload.ai_confidence
-        : null;
-  const nestedTasteVector =
-    coffeePreferences?.taste_vector && typeof coffeePreferences.taste_vector === 'object'
-      ? (coffeePreferences.taste_vector as Record<string, number>)
-      : null;
-  const topLevelTasteVector =
-    payload.taste_vector && typeof payload.taste_vector === 'object'
-      ? (payload.taste_vector as Record<string, number>)
-      : null;
-  if (rawCoffeePreferences && topLevelTasteVector && !nestedTasteVector) {
-    console.warn(
-      'CoffeeTasteScanner: Received legacy top-level taste_vector; expected coffee_preferences.taste_vector.',
-    );
-  }
-  const tasteVector = nestedTasteVector ?? topLevelTasteVector;
-  const updatedAt =
-    typeof payload.updated_at === 'string'
-      ? payload.updated_at
-      : typeof coffeePreferences.updated_at === 'string'
-        ? coffeePreferences.updated_at
-      : typeof payload.updatedAt === 'string'
-        ? payload.updatedAt
-        : null;
-  const lastRecalculatedAt =
-    typeof payload.last_recalculated_at === 'string'
-      ? payload.last_recalculated_at
-      : typeof coffeePreferences.last_recalculated_at === 'string'
-        ? coffeePreferences.last_recalculated_at
-      : typeof payload.lastRecalculatedAt === 'string'
-        ? payload.lastRecalculatedAt
-        : null;
-
-  if (!aiRecommendation && consistencyScore == null && !tasteVector) {
-    return null;
-  }
-
-  return {
-    ai_recommendation: aiRecommendation,
-    consistency_score: consistencyScore,
-    taste_vector: tasteVector,
-    updatedAt,
-    lastRecalculatedAt,
-  };
-};
-
-const normalizeReasonLine = (value: string): string => value.replace(/^•\s*/, '').trim();
-
-const normalizeComparisonText = (value: string): string => {
-  return value
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .toLowerCase()
-    .trim();
-};
-
-const uniqueLines = (lines: string[]): string[] => Array.from(new Set(lines));
-
-const buildComparisonText = (
-  evaluation: CoffeeEvaluationResult | null | undefined,
-  preferenceSnapshot: CoffeePreferenceSnapshot | null | undefined,
-  coffeePreferences: Record<string, unknown> | null | undefined,
-): string => {
-  if (!evaluation && !preferenceSnapshot) {
-    return 'Zatiaľ nemáme dôvody prečo by káva sedela alebo nie.';
-  }
-
-  const verdictExplanation = evaluation?.verdict_explanation;
-  const comparisonSummary =
-    verdictExplanation && typeof verdictExplanation === 'object'
-      && typeof verdictExplanation.comparison_summary === 'string'
-      ? verdictExplanation.comparison_summary.trim()
-      : '';
-  const comparisonSummaryLines = comparisonSummary
-    ? extractSentenceBlocks(comparisonSummary)
-    : [];
-  const normalizedSummaryLines = comparisonSummaryLines.map(line =>
-    normalizeComparisonText(normalizeReasonLine(line)),
-  );
-  const normalizedSummaryText = comparisonSummary
-    ? normalizeComparisonText(comparisonSummary)
-    : '';
-  const shouldOmitDuplicateLine = (line: string): boolean => {
-    if (!normalizedSummaryLines.length) {
-      return false;
-    }
-    const normalizedLine = normalizeComparisonText(normalizeReasonLine(line));
-    if (!normalizedLine) {
-      return false;
-    }
-    return normalizedSummaryLines.some(summaryLine =>
-      summaryLine.includes(normalizedLine) || normalizedLine.includes(summaryLine),
-    );
-  };
-
-  const shouldOmitGenericTip = (line: string): boolean => {
-    const lower = line.toLowerCase();
-    return [
-      'mlet',
-      'grind',
-      'podobn',
-      'podobné kávy',
-      'podobne kavy',
-    ].some(keyword => lower.includes(keyword));
-  };
-
-  const comparison = buildScanPreferenceComparison({
-    evaluation,
-    coffeePreferences: preferenceSnapshot ?? coffeePreferences ?? null,
-    tasteVector: preferenceSnapshot?.taste_vector ?? null,
-  });
-
-  const matchLines = comparison.dimensions
-    .filter(dimension => dimension.match === 'match')
-    .map(dimension => dimension.line)
-    .filter((line): line is string => Boolean(line));
-  const mismatchLines = comparison.dimensions
-    .filter(dimension => dimension.match === 'mismatch')
-    .map(dimension => dimension.line)
-    .filter((line): line is string => Boolean(line));
-  const filteredMatchLines = matchLines.filter(
-    line => !shouldOmitDuplicateLine(line) && !shouldOmitGenericTip(line),
-  );
-  const filteredMismatchLines = mismatchLines.filter(
-    line => !shouldOmitDuplicateLine(line) && !shouldOmitGenericTip(line),
-  );
-  const summaryHasMatchSentence = matchLines.some(line => {
-    const normalizedLine = normalizeComparisonText(normalizeReasonLine(line));
-    if (!normalizedLine || !normalizedSummaryText) {
-      return false;
-    }
-    return (
-      normalizedSummaryText.includes(normalizedLine)
-      || normalizedLine.includes(normalizedSummaryText)
-    );
-  });
-  const supplementalLines = summaryHasMatchSentence
-    ? []
-    : uniqueLines([
-        ...filteredMatchLines,
-        ...filteredMismatchLines,
-      ]).slice(0, 2);
-
-  const primaryText = comparisonSummary
-    ? comparisonSummary
-    : 'Zatiaľ nemáme dôvody prečo by káva sedela alebo nie.';
-
-  return supplementalLines.length
-    ? `${primaryText}\n${supplementalLines.join('\n')}`
-    : primaryText;
-};
-
-const WELCOME_GRADIENT = ['#FF9966', '#A86B8C'];
-const COFFEE_GRADIENT = ['#8B6544', '#6B4423'];
-const WARM_GRADIENT = ['#FFA000', '#FF6B6B'];
-
-const FLAVOR_KEYWORDS = [
-  { keyword: 'kvet', label: '🌺 Kvetinová' },
-  { keyword: 'citr', label: '🍋 Citrusová' },
-  { keyword: 'brosky', label: '🍑 Broskyňa' },
-  { keyword: 'med', label: '🍯 Medová' },
-  { keyword: 'čaj', label: '🍵 Čajová' },
-  { keyword: 'čokol', label: '🍫 Čokoládová' },
-  { keyword: 'karamel', label: '🍮 Karamelová' },
-  { keyword: 'ovoc', label: '🍒 Ovocná' },
-];
-
-const MAX_IMAGE_DIMENSION = 1600;
-const IMAGE_QUALITY = 0.75;
-const MAX_BASE64_LENGTH = 1300000;
-
-type StructuredFieldKey = keyof Pick<
-  StructuredCoffeeMetadata,
-  'roaster' | 'origin' | 'roastLevel' | 'processing' | 'flavorNotes' | 'roastDate' | 'varietals'
->;
-type StructuredTextFieldKey = Exclude<StructuredFieldKey, 'flavorNotes' | 'varietals'>;
-type StructuredListFieldKey = Extract<StructuredFieldKey, 'flavorNotes' | 'varietals'>;
-
-type StructuredFieldState<T> = {
-  value: T;
-  isAutoFilled: boolean;
-  confidence: number | null;
-  warning: string | null;
-};
-
-type StructuredTextFieldState = StructuredFieldState<string | null>;
-type StructuredListFieldState = StructuredFieldState<string[] | null>;
-
-interface StructuredFieldsState {
-  roaster: StructuredTextFieldState;
-  origin: StructuredTextFieldState;
-  roastLevel: StructuredTextFieldState;
-  processing: StructuredTextFieldState;
-  flavorNotes: StructuredListFieldState;
-  roastDate: StructuredTextFieldState;
-  varietals: StructuredListFieldState;
-}
-
-const STRUCTURED_CONFIDENCE_KEYS: Record<StructuredFieldKey, string[]> = {
-  roaster: ['roaster', 'roaster_name', 'brand'],
-  origin: ['origin'],
-  roastLevel: ['roastLevel', 'roast_level'],
-  processing: ['processing', 'process'],
-  flavorNotes: ['flavorNotes', 'flavor_notes', 'notes'],
-  roastDate: ['roastDate', 'roast_date'],
-  varietals: ['varietals', 'variety', 'varieties'],
-};
-
-const STRUCTURED_FIELD_LABELS: Record<StructuredFieldKey, string> = {
-  roaster: 'Pražiareň',
-  origin: 'Pôvod',
-  roastLevel: 'Stupeň praženia',
-  processing: 'Spracovanie',
-  flavorNotes: 'Chuťové tóny',
-  roastDate: 'Dátum praženia',
-  varietals: 'Odrody',
-};
-
-const resolveVerdictExplanationText = (
-  payload: CoffeeEvaluationResult['verdict_explanation'] | null | undefined,
-): string => {
-  if (!payload) {
-    return '';
-  }
-  if (typeof payload === 'string') {
-    return payload.trim();
-  }
-  return (
-    payload.comparison_summary
-    || payload.user_preferences_summary
-    || payload.coffee_profile_summary
-    || ''
-  ).trim();
-};
-
-const extractSentenceBlocks = (text: string): string[] => {
-  if (!text) {
-    return [];
-  }
-  const parsed = parseAIResponse(text);
-  const parsedLines = parsed.sections.flatMap(section => [
-    ...(section.content ?? []),
-    ...(section.bullets ?? []),
-  ]);
-  const sourceLines = parsedLines.length ? parsedLines : text.split('\n');
-  const sentences = sourceLines.flatMap(line => line.split(/[.!?]+/));
-  return uniqueLines(sentences.map(sentence => sentence.trim()).filter(Boolean));
-};
-
-const resolveInsightHeadline = (insight: CoffeeEvaluationResult['insight'] | null | undefined): string => {
-  if (!insight || typeof insight.headline !== 'string') {
-    return '';
-  }
-  return insight.headline.trim();
-};
-
-const STRUCTURED_FIELD_ORDER: Array<{
-  key: StructuredFieldKey;
-  type: 'text' | 'list';
-  placeholder: string;
-}> = [
-  { key: 'roaster', type: 'text', placeholder: 'Napr. pražiareň alebo brand' },
-  { key: 'origin', type: 'text', placeholder: 'Napr. krajina alebo región pôvodu' },
-  { key: 'roastLevel', type: 'text', placeholder: 'Napr. light, medium, dark' },
-  { key: 'processing', type: 'text', placeholder: 'Napr. washed, natural, honey' },
-  { key: 'flavorNotes', type: 'list', placeholder: 'Oddeluj čiarkou: čokoláda, čerešňa' },
-  { key: 'roastDate', type: 'text', placeholder: 'Napr. 12.02.2024' },
-  { key: 'varietals', type: 'list', placeholder: 'Oddeluj čiarkou: Bourbon, Typica' },
-];
-
-const normalizeStructuredStringValue = (value: string | null | undefined): string | null => {
-  if (typeof value !== 'string') {
-    return null;
-  }
-  const trimmed = value.trim();
-  return trimmed.length > 0 ? trimmed : null;
-};
-
-const normalizeStructuredStringArrayValue = (
-  value: string[] | string | null | undefined,
-): string[] | null => {
-  if (Array.isArray(value)) {
-    const cleaned = value
-      .map(item => (typeof item === 'string' ? item.trim() : ''))
-      .filter(Boolean);
-    return cleaned.length ? cleaned : null;
-  }
-
-  if (typeof value === 'string') {
-    const parts = value
-      .split(/[,;\n]/)
-      .map(part => part.trim())
-      .filter(Boolean);
-    return parts.length ? parts : null;
-  }
-
-  return null;
-};
-
-const isStructuredValueFilled = <T,>(value: T): boolean => {
-  if (Array.isArray(value)) {
-    return value.length > 0;
-  }
-  if (typeof value === 'string') {
-    return value.trim().length > 0;
-  }
-  return value != null;
-};
-
-const resolveConfidenceEntry = (
-  source: Record<string, unknown> | null | undefined,
-  key: StructuredFieldKey,
-  fallbackFlag?: boolean | null,
-): { confidence: number | null; warning: string | null } => {
-  const booleanConfidenceEstimate = 0.8;
-  const lowConfidenceThreshold = 0.85;
-  const aliases = STRUCTURED_CONFIDENCE_KEYS[key] ?? [key];
-  let confidence: number | null = null;
-  let warning: string | null = null;
-
-  const assignFromValue = (entry: unknown) => {
-    if (typeof entry === 'number') {
-      confidence = entry;
-      return;
-    }
-    if (typeof entry === 'boolean') {
-      confidence = entry ? booleanConfidenceEstimate : null;
-      return;
-    }
-    if (typeof entry === 'string') {
-      const trimmed = entry.trim();
-      if (!trimmed) {
-        return;
-      }
-      const normalized = Number.parseFloat(trimmed.replace(',', '.'));
-      if (!Number.isNaN(normalized)) {
-        confidence = normalized;
-        return;
-      }
-      warning = trimmed;
-      return;
-    }
-    if (entry && typeof entry === 'object') {
-      const container = entry as Record<string, unknown>;
-      if (container.confidence != null) {
-        assignFromValue(container.confidence);
-      } else if (container.score != null) {
-        assignFromValue(container.score);
-      } else if (container.value != null) {
-        assignFromValue(container.value);
-      }
-      if (warning == null && typeof container.warning === 'string') {
-        warning = container.warning;
-      }
-    }
-  };
-
-  if (source) {
-    for (const alias of aliases) {
-      if (Object.prototype.hasOwnProperty.call(source, alias)) {
-        assignFromValue(source[alias]);
-        break;
-      }
-    }
-  }
-
-  if (confidence == null && typeof fallbackFlag === 'boolean') {
-    confidence = fallbackFlag ? booleanConfidenceEstimate : null;
-  }
-
-  if (confidence != null) {
-    const normalized = confidence > 1 ? confidence / 100 : confidence;
-    if (normalized < lowConfidenceThreshold && warning == null) {
-      warning = 'AI si nie je isté';
-    }
-  }
-
-  return { confidence, warning };
-};
-
-const createStructuredFieldsFromMetadata = (
-  metadata: StructuredCoffeeMetadata | null | undefined,
-  confidence: Record<string, unknown> | null | undefined,
-): StructuredFieldsState => {
-  const buildTextField = (
-    key: StructuredTextFieldKey,
-    value: string | null | undefined,
-  ): StructuredTextFieldState => {
-    const normalized = normalizeStructuredStringValue(value);
-    const flag = metadata?.confidenceFlags?.[key] ?? null;
-    const { confidence: parsedConfidence, warning } = resolveConfidenceEntry(confidence, key, flag);
-    return {
-      value: normalized,
-      isAutoFilled: isStructuredValueFilled(normalized),
-      confidence: parsedConfidence,
-      warning,
-    };
-  };
-
-  const buildListField = (
-    key: StructuredListFieldKey,
-    value: string[] | null | undefined,
-  ): StructuredListFieldState => {
-    const normalized = normalizeStructuredStringArrayValue(value);
-    const flag = metadata?.confidenceFlags?.[key] ?? null;
-    const { confidence: parsedConfidence, warning } = resolveConfidenceEntry(confidence, key, flag);
-    return {
-      value: normalized,
-      isAutoFilled: isStructuredValueFilled(normalized),
-      confidence: parsedConfidence,
-      warning,
-    };
-  };
-
-  return {
-    roaster: buildTextField('roaster', metadata?.roaster),
-    origin: buildTextField('origin', metadata?.origin),
-    roastLevel: buildTextField('roastLevel', metadata?.roastLevel),
-    processing: buildTextField('processing', metadata?.processing),
-    flavorNotes: buildListField('flavorNotes', metadata?.flavorNotes ?? null),
-    roastDate: buildTextField('roastDate', metadata?.roastDate),
-    varietals: buildListField('varietals', metadata?.varietals ?? null),
-  };
-};
-
-const structuredFieldsToMetadata = (
-  fields: StructuredFieldsState,
-): StructuredCoffeeMetadata | null => {
-  const metadata: StructuredCoffeeMetadata = {
-    roaster: normalizeStructuredStringValue(fields.roaster.value),
-    origin: normalizeStructuredStringValue(fields.origin.value),
-    roastLevel: normalizeStructuredStringValue(fields.roastLevel.value),
-    processing: normalizeStructuredStringValue(fields.processing.value),
-    flavorNotes: normalizeStructuredStringArrayValue(fields.flavorNotes.value) ?? null,
-    roastDate: normalizeStructuredStringValue(fields.roastDate.value),
-    varietals: normalizeStructuredStringArrayValue(fields.varietals.value) ?? null,
-    confidenceFlags: null,
-  };
-
-  const hasValues = [
-    metadata.roaster,
-    metadata.origin,
-    metadata.roastLevel,
-    metadata.processing,
-    metadata.roastDate,
-  ].some(Boolean) ||
-    (metadata.flavorNotes?.length ?? 0) > 0 ||
-    (metadata.varietals?.length ?? 0) > 0;
-
-  if (!hasValues) {
-    return null;
-  }
-
-  const confidenceFlags: StructuredCoffeeMetadata['confidenceFlags'] = {};
-
-  (['roaster', 'origin', 'roastLevel', 'processing', 'flavorNotes', 'roastDate', 'varietals'] as StructuredFieldKey[])
-    .forEach(key => {
-      const field = fields[key];
-      const hasValue = isStructuredValueFilled(field.value);
-      if (!hasValue) {
-        return;
-      }
-      confidenceFlags[key] = field.isAutoFilled ? true : false;
-    });
-
-  metadata.confidenceFlags = Object.keys(confidenceFlags).length > 0 ? confidenceFlags : null;
-
-  return metadata;
-};
-
-const structuredFieldsToConfidence = (
-  fields: StructuredFieldsState,
-): Record<string, unknown> | null => {
-  const result: Record<string, unknown> = {};
-
-  (['roaster', 'origin', 'roastLevel', 'processing', 'flavorNotes', 'roastDate', 'varietals'] as StructuredFieldKey[])
-    .forEach(key => {
-      const field = fields[key];
-      const payload: Record<string, unknown> = {};
-
-      if (field.confidence != null) {
-        payload.confidence = field.confidence;
-      }
-      if (!field.isAutoFilled) {
-        payload.isManual = true;
-      }
-      if (field.warning) {
-        payload.warning = field.warning;
-      }
-
-      if (Object.keys(payload).length > 0) {
-        result[key] = payload;
-      }
-    });
-
-  return Object.keys(result).length > 0 ? result : null;
-};
-
-const clampTasteValue = (value: number): number => {
-  if (Number.isNaN(value)) {
-    return 5;
-  }
-  return Math.min(10, Math.max(0, value));
-};
-
-const normalizeTasteVectorTo10 = (
-  value?: Record<string, number> | null,
-): TasteProfileVector | null => {
-  if (!value) {
-    return null;
-  }
-
-  const parseValue = (entry: unknown): number | null => {
-    if (typeof entry === 'number' && Number.isFinite(entry)) {
-      return clampTasteValue(entry <= 1 ? entry * 10 : entry);
-    }
-    if (typeof entry === 'string') {
-      const parsed = Number(entry);
-      if (Number.isFinite(parsed)) {
-        return clampTasteValue(parsed <= 1 ? parsed * 10 : parsed);
-      }
-    }
-    return null;
-  };
-
-  const sweetness = parseValue(value.sweetness);
-  const acidity = parseValue(value.acidity);
-  const bitterness = parseValue(value.bitterness);
-  const body = parseValue(value.body);
-
-  if (![sweetness, acidity, bitterness, body].some(entry => entry !== null)) {
-    return null;
-  }
-
-  return {
-    sweetness: sweetness ?? 5,
-    acidity: acidity ?? 5,
-    bitterness: bitterness ?? 5,
-    body: body ?? 5,
-  };
-};
-
-const extractTasteVectorFromPayload = (payload: unknown): TasteProfileVector | null => {
-  if (!payload || typeof payload !== 'object') {
-    return null;
-  }
-
-  const record = payload as Record<string, unknown>;
-  const nestedTasteVector = (input: unknown): Record<string, number> | null => {
-    if (!input || typeof input !== 'object') {
-      return null;
-    }
-    return input as Record<string, number>;
-  };
-
-  const directCandidate =
-    nestedTasteVector(record.taste_vector) ??
-    nestedTasteVector(record.tasteVector) ??
-    nestedTasteVector(record.taste_profile) ??
-    nestedTasteVector(record.tasteProfile) ??
-    nestedTasteVector(
-      typeof record.taste_profile === 'object'
-        ? (record.taste_profile as Record<string, unknown>).taste_vector
-        : null,
-    ) ??
-    nestedTasteVector(
-      typeof record.tasteProfile === 'object'
-        ? (record.tasteProfile as Record<string, unknown>).taste_vector ??
-          (record.tasteProfile as Record<string, unknown>).tasteVector
-        : null,
-    ) ??
-    nestedTasteVector(
-      typeof record.coffee_attributes === 'object'
-        ? (record.coffee_attributes as Record<string, unknown>).taste_vector
-        : null,
-    ) ??
-    nestedTasteVector(
-      typeof record.coffeeAttributes === 'object'
-        ? (record.coffeeAttributes as Record<string, unknown>).taste_vector ??
-          (record.coffeeAttributes as Record<string, unknown>).tasteVector
-        : null,
-    );
-
-  return directCandidate ? normalizeTasteVectorTo10(directCandidate) : null;
-};
-
-const hasCompleteTasteVector = (vector?: TasteProfileVector | null): boolean => {
-  if (!vector) {
-    return false;
-  }
-  return ['sweetness', 'acidity', 'bitterness', 'body'].every(key => {
-    const value = vector[key as keyof TasteProfileVector];
-    return typeof value === 'number' && Number.isFinite(value);
-  });
-};
-
-const resolveTasteProfile = (
-  profile?: TasteProfileVector | UserTasteProfile | null,
-  snapshotVector?: TasteProfileVector | null,
-): TasteProfileVector | UserTasteProfile | null => {
-  if (profile && 'preferences' in profile) {
-    if (hasCompleteTasteVector(profile.preferences ?? null)) {
-      return profile;
-    }
-  } else if (hasCompleteTasteVector(profile ?? null)) {
-    return profile ?? null;
-  }
-
-  if (hasCompleteTasteVector(snapshotVector ?? null)) {
-    return snapshotVector ?? null;
-  }
-
-  return null;
-};
-
 type CompatibilityBucket = 'SAFE' | 'RISKY' | 'NO-GO';
-
-const normalizeConfidenceScore = (confidence?: number | null): number | null => {
-  if (typeof confidence !== 'number') {
-    return null;
-  }
-  const normalized = confidence <= 1 ? confidence * 100 : confidence;
-  return Math.round(Math.max(0, Math.min(100, normalized)));
-};
-
 
 const CoffeeTasteScanner: React.FC<ProfessionalOCRScannerProps> = ({
   onBack,
